final changes - removed extra files
