+++ web/templates/success_stories/create.html
@@ -80,7 +80,7 @@
         </div>
 
         <!-- Status -->
-        <div>
+        <div class="hidden">
           <label for="{{ form.status.id_for_label }}" class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
             Status
           </label>
